EvolveCore/pyproject.toml
[project]
name = "flash-searcher"
version = "0.1.0"
description = "Flash-Searcher: Fast and Effective Web Agents via DAG-Based Parallel Execution"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "OPPO PersonalAI Team"}
]
keywords = ["agent", "dag", "web-search", "parallel-execution"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # AI/ML 核心框架 (代码中实际使用)
    "accelerate>=1.5.2",
    "anthropic>=0.49.0",
    "datasets>=3.4.1",
    "huggingface-hub>=0.29.3",
    "langchain>=0.3.23",
    "litellm>=1.63.12",
    "openai>=1.74.0",
    "sentence_transformers>=4.1.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.49.0",
    
    # 多模态处理 (代码中直接导入)
    "beautifulsoup4>=4.13.3",
    "biopython>=1.85",
    "mammoth>=1.9.0",
    "markdownify>=1.1.0",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdfminer.six>=20240706",
    "pillow>=11.1.0",
    "pydub>=0.25.1",
    "python-pptx>=1.0.2",
    "puremagic>=1.30",
    "SpeechRecognition>=3.14.1",
    "youtube-transcript-api>=1.0.2",
    
    # Web 框架和工具
    "fastapi>=0.115.11",
    "Flask>=3.1.0",
    "flask-cors>=5.0.1",
    "uvicorn>=0.34.0",
    
    # 搜索和爬虫工具
    "apify_client>=1.10.0",
    "crawl4ai>=0.7.3",
    "duckduckgo_search>=7.5.2",
    "google_search_results>=2.4.2",
    "serpapi>=0.1.5",
    
    # 监控和可观测性
    "arize-phoenix>=8.16.0",
    
    # MCP 和 E2B 工具
    "e2b>=1.2.0",
    "e2b-code-interpreter>=1.1.1",
    "mcp>=1.4.1",
    "mcpadapt>=0.0.15",
    
    # 数据库
    "SQLAlchemy>=2.0.39",
    "sqlean.py>=3.47.0",
    
    # 核心工具库 (代码中直接导入)
    "jinja2>=3.1.6",
    "json_repair>=0.44.1",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["FlashOAgents"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.1",
]
